+++ ChangeLog.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
-## [1.7.12] - 2020-01-xx
+## [1.7.12] - 2020-01-03
 
 ### Added